Added events handler
