+++ events/ready.js
@@ -1,0 +1,5 @@
+module.exports = {
+	execute(client) {
+		console.log(`${client.user.tag} is ready.`);
+	},
+};
