+++ util/util.js
@@ -1,4 +1,5 @@
 const fs = require('fs');
+const path = require('path');
 
 const convertUserRolesToArray = (userRoles) => {
 	const userRoleIds = [];
@@ -12,7 +13,7 @@ const checkPermission = (userRoles, authorizedRoles) => {
 	return userRoles.some(userRole => authorizedRoles.includes(userRole));
 };
 
-const retriveAuthroles = async (guildId, DB) => {
+const retriveAuthroles = async (message, guildId, DB) => {
 	try {
 		const snap = await DB.collection(`Server: ${guildId}`).doc(`Data for server: ${guildId}`).get();
 		if (!snap.exists) {
@@ -26,7 +27,7 @@ const retriveAuthroles = async (guildId, DB) => {
 		return authorizedRoles;
 	}
 	catch (error) {
-		throw (`${error}`);
+		return message.channel.send(`${error}`);
 	}
 
 };
@@ -37,19 +38,23 @@ const loadCommands = (client) => {
 	for (const folder of commandFolder) {
 		const commandFiles = fs.readdirSync(`../commands/${folder}`).filter(file => file.endsWith('.js'));
 		for (const file of commandFiles) {
-			const command = require(`./commands/${folder}/${file}`);
+			const command = require(`../commands/${folder}/${file}`);
 			client.commands.set(command.name, command);
 		}
 	}
-	console.log('Loaded.');
+	console.log('Loaded all commands.');
 };
 
 const loadEvents = (client) => {
 	const eventsFolder = fs.readdirSync('../events').filter(file => file.endsWith('.js'));
+
 	for (const events of eventsFolder) {
+		const { name:eventName } = path.parse(events);
 		const event = require(`../events/${events}`);
-		client.on(event, event.bind(null, client));
+		client.on(eventName, event.bind(null, client));
+		delete require.cache[event];
 	}
+	console.log('Loaded all events');
 };
 
 module.exports = {
